+++ apps/backend/src/app/api/v1/users/crud.tsx
@@ -80,13 +80,15 @@ export const userPrismaToCrud = (
     id: prisma.projectUserId,
     display_name: prisma.displayName || null,
     primary_email: primaryEmailContactChannel?.value || null,
-    primary_email_verified: primaryEmailContactChannel?.isVerified || false,
+    primary_email_verified: !!primaryEmailContactChannel?.isVerified,
+    primary_email_auth_enabled: !!primaryEmailContactChannel?.usedForAuth,
     profile_image_url: prisma.profileImageUrl,
     signed_up_at_millis: prisma.createdAt.getTime(),
     client_metadata: prisma.clientMetadata,
     client_read_only_metadata: prisma.clientReadOnlyMetadata,
     server_metadata: prisma.serverMetadata,
     has_password: !!passwordAuth,
+    otp_auth_enabled: !!otpAuth,
     auth_with_email: !!passwordAuth || !!otpAuth,
     requires_totp_mfa: prisma.requiresTotpMfa,
     oauth_providers: prisma.projectUserOAuthAccounts.map((a) => ({
@@ -117,9 +119,6 @@ async function checkAuthData(
   if (!data.primaryEmail && data.primaryEmailVerified) {
     throw new StatusError(400, "primary_email_verified cannot be true without primary_email");
   }
-  if (!data.primaryEmailAuthEnabled && data.passwordHash) {
-    throw new StatusError(400, "password cannot be set without primary_email_auth_enabled");
-  }
   if (data.primaryEmailAuthEnabled) {
     if (!data.oldPrimaryEmail || data.oldPrimaryEmail !== data.primaryEmail) {
       const otpAuth = await tx.contactChannel.findFirst({
@@ -142,7 +141,10 @@ async function checkAuthData(
 async function getPasswordConfig(tx: PrismaTransaction, projectConfigId: string) {
   const passwordConfig = await tx.passwordAuthMethodConfig.findMany({
     where: {
-      projectConfigId: projectConfigId
+      projectConfigId: projectConfigId,
+      authMethodConfig: {
+        enabled: true,
+      }
     },
     include: {
       authMethodConfig: true,
@@ -160,7 +162,10 @@ async function getPasswordConfig(tx: PrismaTransaction, projectConfigId: string)
 async function getOtpConfig(tx: PrismaTransaction, projectConfigId: string) {
   const otpConfig = await tx.otpAuthMethodConfig.findMany({
     where: {
-      projectConfigId: projectConfigId
+      projectConfigId: projectConfigId,
+      authMethodConfig: {
+        enabled: true,
+      }
     },
     include: {
       authMethodConfig: true,
@@ -374,55 +379,57 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
             projectUserId: newUser.projectUserId,
             projectId: auth.project.id,
             type: 'EMAIL' as const,
-            value: data.primary_email || throwErr("primary_email_auth_enabled is true but primary_email is not set"),
+            value: data.primary_email,
             isVerified: data.primary_email_verified ?? false,
             isPrimary: "TRUE",
             usedForAuth: data.primary_email_auth_enabled ? BooleanTrue.TRUE : null,
           }
         });
+      }
 
-        if (data.primary_email_auth_enabled) {
-          const otpConfig = await getOtpConfig(tx, auth.project.config.id);
+      if (data.password) {
+        const passwordConfig = await getPasswordConfig(tx, auth.project.config.id);
 
-          if (otpConfig) {
-            await tx.authMethod.create({
-              data: {
-                projectId: auth.project.id,
-                projectUserId: newUser.projectUserId,
-                projectConfigId: auth.project.config.id,
-                authMethodConfigId: otpConfig.authMethodConfigId,
-                otpAuthMethod: {
-                  create: {
-                    projectUserId: newUser.projectUserId,
-                  }
-                }
-              }
-            });
-          }
+        if (!passwordConfig) {
+          throw new StatusError(StatusError.BadRequest, "Password auth not enabled in the project");
         }
 
-        if (data.password) {
-          const passwordConfig = await getPasswordConfig(tx, auth.project.config.id);
-
-          if (!passwordConfig) {
-            throw new StatusError(StatusError.BadRequest, "Password auth not enabled in the project");
-          }
-
-          await tx.authMethod.create({
-            data: {
-              projectId: auth.project.id,
-              projectConfigId: auth.project.config.id,
-              projectUserId: newUser.projectUserId,
-              authMethodConfigId: passwordConfig.authMethodConfigId,
-              passwordAuthMethod: {
-                create: {
-                  passwordHash: await hashPassword(data.password),
-                  projectUserId: newUser.projectUserId,
-                }
+        await tx.authMethod.create({
+          data: {
+            projectId: auth.project.id,
+            projectConfigId: auth.project.config.id,
+            projectUserId: newUser.projectUserId,
+            authMethodConfigId: passwordConfig.authMethodConfigId,
+            passwordAuthMethod: {
+              create: {
+                passwordHash: await hashPassword(data.password),
+                projectUserId: newUser.projectUserId,
               }
             }
-          });
+          }
+        });
+      }
+
+      if (data.otp_auth_enabled) {
+        const otpConfig = await getOtpConfig(tx, auth.project.config.id);
+
+        if (!otpConfig) {
+          throw new StatusError(StatusError.BadRequest, "OTP auth not enabled in the project");
         }
+
+        await tx.authMethod.create({
+          data: {
+            projectId: auth.project.id,
+            projectConfigId: auth.project.config.id,
+            projectUserId: newUser.projectUserId,
+            authMethodConfigId: otpConfig.authMethodConfigId,
+            otpAuthMethod: {
+              create: {
+                projectUserId: newUser.projectUserId,
+              }
+            }
+          }
+        });
       }
 
       const user = await tx.projectUser.findUnique({
@@ -527,7 +534,7 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
         oldPrimaryEmail: primaryEmailContactChannel?.value,
         primaryEmail: primaryEmailContactChannel?.value || data.primary_email,
         primaryEmailVerified: primaryEmailContactChannel?.isVerified || data.primary_email_verified,
-        primaryEmailAuthEnabled: !!otpAuth || !!passwordAuth || data.primary_email_auth_enabled,
+        primaryEmailAuthEnabled: !!primaryEmailContactChannel?.usedForAuth || data.primary_email_auth_enabled,
         passwordHash: passwordAuth ? passwordAuth.passwordHash : (data.password && await hashPassword(data.password)),
       });
 
@@ -592,26 +599,13 @@ export const usersCrudHandlers = createLazyProxy(() => createCrudHandlers(usersC
         });
       }
 
-      // if primary email auth is true
+      // if otp_auth_enabled is true
       // - create a new otp auth method if it doesn't exist
-      // if primary email auth is false
+      // if otp_auth_enabled is false
       // - delete the otp auth method if it exists
-      if (data.primary_email_auth_enabled !== undefined) {
-        if (data.primary_email_auth_enabled) {
+      if (data.otp_auth_enabled !== undefined) {
+        if (data.otp_auth_enabled) {
           if (!otpAuth) {
-            const primaryEmailChannel = await tx.contactChannel.findFirst({
-              where: {
-                projectId: auth.project.id,
-                projectUserId: params.user_id,
-                type: 'EMAIL',
-                isPrimary: "TRUE",
-              }
-            });
-
-            if (!primaryEmailChannel) {
-              throw new StackAssertionError("primary_email_auth_enabled is true but primary_email is not set");
-            }
-
             const otpConfig = await getOtpConfig(tx, auth.project.config.id);
 
             if (otpConfig) {