+++ examples/cjs-test/CHANGELOG.md
@@ -1,4 +1,19 @@
-# cjs-test
+# @stackframe/example-cjs-test
+
+## 2.6.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated account settings page
+- Updated dependencies
+  - @stackframe/stack@2.6.12
 
 ## 2.6.11